updating label spec tests
