+++ app/views/labels/_form.html.erb
@@ -12,8 +12,7 @@
   <% end %>
 
   <div class="field">
-    <%= form.label :name %>
-    <%= form.text_field :name %>
+    <%= form.text_field :name, placeholder: 'name' %>
   </div>
 
   <div class="actions">
