+++ spec/system/labels_spec.rb
@@ -17,7 +17,7 @@ RSpec.describe "Labels function", type: :system do
         visit new_session_path
         fill_in 'session_email', with: 'user1@example.com'
         fill_in 'session_password', with: 'password'
-        click_on 'Log in'
+        click_button 'Login'
       end
       it 'should have access to label screen' do
         visit labels_path
@@ -56,7 +56,7 @@ RSpec.describe "Labels function", type: :system do
       visit new_session_path
       fill_in 'session_email', with: 'user1@example.com'
       fill_in 'session_password', with: 'password'
-      click_on 'Log in'
+      click_button 'Login'
     end
 
     context 'when new task is created' do
@@ -112,7 +112,9 @@ RSpec.describe "Labels function", type: :system do
         click_on 'Submit'
       end
       it 'should remove label' do
-        click_link 'Edit', match: :first
+        within first('tbody tr') do
+          click_on 'Edit'
+         end
         uncheck 'SecondLabel'
         uncheck 'ThirdLabel'
         click_on 'Submit'
@@ -121,22 +123,123 @@ RSpec.describe "Labels function", type: :system do
         expect(page).to have_no_content 'ThirdLabel'
       end
     end
+  end
 
-    context 'searching with label' do
-      before do
-        visit tasks_path
-        click_link 'Edit', match: :first
-        check 'ThirdLabel'
+  describe 'related to task and labels' do
+    before do
+      visit new_session_path
+      fill_in 'session_email', with: 'user1@example.com'
+      fill_in 'session_password', with: 'password'
+      click_button 'Login'
+    end
+
+    context 'when new task is created' do
+      it 'should be able to add a label' do
+        visit new_task_path
+        fill_in "Task Name", with: 'title test'
+        fill_in "Task Details", with: 'content test'
+        select 'started'
+        select 'low'
+        check 'label'
+        check 'second_label'
+        check 'third_task'
         click_on 'Submit'
-      end
-      it 'only task with selected label are displayed' do
-        visit tasks_path
-        check "search_label_ids_609"
-        click_on 'search'
-        expect(page).to have_no_content 'future'
-        expect(page).to have_no_content 'old'
+        expect(page).to have_content 'Task created'
+        expect(page).to have_content 'FirstLabel'
+        expect(page).to have_content 'SecondLabel'
+        expect(page).to have_content 'ThirdLabel'
       end
     end
 
+    context 'when visiting task details screen' do
+      it 'should display the label' do
+        visit task_path(@task.id)
+        expect(page).to have_content 'test1'
+        expect(page).not_to have_content 'test2'
+        expect(page).to have_content 'FirstLabel'
+        expect(page).to have_content 'SecondLabel'
+        expect(page).not_to have_content 'ThirdLabel'
+      end
+    end
+
+    context 'editing labels' do
+      it 'should be able to update labels' do
+        visit tasks_path
+        within first('tbody tr') do
+          click_on 'Edit'
+         end
+        check 'FirstLabel'
+        check 'SecondLabel'
+        check 'ThirdLabel'
+        click_on 'Submit'
+        within first('tbody tr') do
+          click_on 'Show'
+         end
+        expect(page).to have_content 'FirstLabel'
+        expect(page).to have_content 'SecondLabel'
+        expect(page).to have_content 'ThirdLabel'
+      end
+    end
+
+    context 'Using edit to delete label' do
+      before do
+        visit tasks_path
+        within first('tbody tr') do
+          click_on 'Edit'
+         end
+        check 'FirstLabel'
+        check 'SecondLabel'
+        check 'ThirdLabel'
+        click_on 'Submit'
+      end
+      it 'should remove label' do
+        within first('tbody tr') do
+          click_on 'Edit'
+         end
+        uncheck 'SecondLabel'
+        uncheck 'ThirdLabel'
+        click_on 'Submit'
+        within first('tbody tr') do
+          click_on 'Show'
+         end
+        expect(page).to have_no_content 'SecondLabel'
+        expect(page).to have_no_content 'ThirdLabel'
+      end
+    end
+
+    describe 'search function' do
+      context 'If you do a fuzzy search for the title, label and status' do
+        it "Narrows down tasks that include search with title label and status" do
+          visit tasks_path
+          fill_in "title keyword", with: "test"
+          check "started"
+          check "MyString02"
+          click_button "search"
+          expect(page).to have_content 'test'
+          expect(page).to have_content 'started'
+          expect(page).to have_content 'MyString02'
+        end
+      end
+
+      context 'If  you  do a fuzzy label search with title' do
+        it "Narrow down search to exactly match the label and title" do
+          visit tasks_path
+          fill_in "title keyword", with: "test"
+          select "MyString01"
+          click_button "search"
+          expect(page).to have_content 'test'
+          expect(page).to have_content 'MyString01'
+        end
+      end
+
+      context 'When a status search is performed with the scope method' do
+        it "Narrows down tasks that exactly match with label" do
+          visit tasks_path
+          check "MyString02"
+          click_button "search"
+          expect(page).to have_content 'MyString02'
+        end
+      end
+    end
   end
 end
